requirements/base.txt
channels==3.0.4
Django==3.2.8
django-anymail==8.4
django-bleach==1.0.0
django-environ==0.7.0
django-hosts==4.0
django-ratelimit==3.0.1
emoji==1.6.1
importlib-resources==5.4.0; python_version < '3.9'
more-itertools==8.12.0
requests==2.26.0
typing-extensions==3.10.0.2
Unidecode==1.3.2
whitenoise==5.3.0
json5==0.9.6
